Apply clippy fixes

+++ adder-codec-rs/src/framer/scale_intensity.rs
@@ -124,16 +124,16 @@ impl FrameValue for u16 {
                 let intensity = event_to_intensity(event);
                 match source_type {
                     SourceType::U8 => {
-                        (intensity / f64::from(u8::MAX) * f64::from(tpf) * f64::from(u16::MAX))
+                        (intensity / f64::from(u8::MAX) * tpf * f64::from(u16::MAX))
                             as u16
                     }
-                    SourceType::U16 => (intensity * f64::from(tpf)) as u16,
+                    SourceType::U16 => (intensity * tpf) as u16,
                     SourceType::U32 => {
-                        (intensity / f64::from(u32::MAX) * f64::from(tpf) * f64::from(u16::MAX))
+                        (intensity / f64::from(u32::MAX) * tpf * f64::from(u16::MAX))
                             as u16
                     }
                     SourceType::U64 => {
-                        (intensity / u64::MAX as f64 * f64::from(tpf) * f64::from(u16::MAX)) as u16
+                        (intensity / u64::MAX as f64 * tpf * f64::from(u16::MAX)) as u16
                     }
                     SourceType::F32 => {
                         todo!()
@@ -177,16 +177,16 @@ impl FrameValue for u32 {
                 let intensity = event_to_intensity(event);
                 match source_type {
                     SourceType::U8 => {
-                        (intensity / f64::from(u8::MAX) * f64::from(tpf) * f64::from(u32::MAX))
+                        (intensity / f64::from(u8::MAX) * tpf * f64::from(u32::MAX))
                             as u32
                     }
                     SourceType::U16 => {
-                        (intensity / f64::from(u16::MAX) * f64::from(tpf) * f64::from(u32::MAX))
+                        (intensity / f64::from(u16::MAX) * tpf * f64::from(u32::MAX))
                             as u32
                     }
-                    SourceType::U32 => (intensity * f64::from(tpf)) as u32,
+                    SourceType::U32 => (intensity * tpf) as u32,
                     SourceType::U64 => {
-                        (intensity / u64::MAX as f64 * f64::from(tpf) * f64::from(u32::MAX)) as u32
+                        (intensity / u64::MAX as f64 * tpf * f64::from(u32::MAX)) as u32
                     }
                     SourceType::F32 => {
                         todo!()
@@ -228,15 +228,15 @@ impl FrameValue for u64 {
                 let intensity = event_to_intensity(event);
                 match source_type {
                     SourceType::U8 => {
-                        (intensity / f64::from(u8::MAX) * f64::from(tpf) * u64::MAX as f64) as u64
+                        (intensity / f64::from(u8::MAX) * tpf * u64::MAX as f64) as u64
                     }
                     SourceType::U16 => {
-                        (intensity / f64::from(u16::MAX) * f64::from(tpf) * u64::MAX as f64) as u64
+                        (intensity / f64::from(u16::MAX) * tpf * u64::MAX as f64) as u64
                     }
                     SourceType::U32 => {
-                        (intensity / f64::from(u32::MAX) * f64::from(tpf) * u64::MAX as f64) as u64
+                        (intensity / f64::from(u32::MAX) * tpf * u64::MAX as f64) as u64
                     }
-                    SourceType::U64 => (intensity * f64::from(tpf)) as u64,
+                    SourceType::U64 => (intensity * tpf) as u64,
                     SourceType::F32 => {
                         todo!()
                     }
